Move LS related collab code to excal app
